move MECS functions out of temp_data_source_functions
