+++ flowsa/data_source_scripts/EIA_MECS.py
@@ -436,6 +436,155 @@ def eia_mecs_energy_parse(*, df_list, source, year, **_):
     return df
 
 
+def estimate_suppressed_mecs_energy(
+        fba: FlowByActivity,
+        **kwargs
+    ) -> FlowByActivity:
+    '''
+    Rough first pass at an estimation method, for testing purposes. This
+    will drop rows with 'D' or 'Q' values, on the grounds that as far as I can
+    tell we don't have any more information for them than we do for any
+    industry without its own line item in the MECS anyway. '*' is for value
+    less than 0.5 Trillion Btu and will be assumed to be 0.25 Trillion Btu
+    '''
+    if 'Suppressed' not in fba.columns:
+        log.warning('The current MECS dataframe does not contain data '
+                    'on estimation method and so suppressed data will '
+                    'not be assessed.')
+        return fba
+    dropped = fba.query('Suppressed not in ["D", "Q"]')
+    unsuppressed = dropped.assign(
+        FlowAmount=dropped.FlowAmount.mask(dropped.Suppressed == '*', 0.25)
+    )
+
+    return unsuppressed.drop(columns='Suppressed')
+
+def clean_mecs_energy_fba(fba: FlowByActivity, **_) -> FlowByActivity:
+    '''
+    This function helps address the unique structure of the EIA MECS dataset.
+    The MECS dataset contains rows at various levels of aggregation between
+    NAICS-3 and NAICS-6 (inclusive). Each aggregated row contains the total
+    for that level of aggregation, even if data are also reported for a less
+    aggregated subset of those industries. For example:
+
+    ActivityConsumedBy | FlowAmount | ...
+    -------------------------------------
+    311                | 110        |
+    3112               |  65        |
+    311221             |  55        |
+
+    where the 110 reported for 311 includes the 65 reported for 3112, which
+    includes the 55 reported for 211221. If we do not address this issue, there
+    will be double counting. Additionally, if we are trying to disaggregate
+    to the NAICS-6 level, all three rows shown above will be mapped to NAICS-6
+    311221 (with the first wo rows also being mapped to several other NAICS-6
+    codes as well). We will then over attribute the (double-counted) flows to
+    those industries and groups of industries for which more specific detail
+    is provided.
+
+    This function addresses the double counting issue. For each aggregated
+    industry group, all descendant (less aggregated) industries or industry
+    groups for which detailed information is given are subtracted from the
+    aggregated total. Using the example from above:
+
+    ActivityConsumedBy | FlowAmount | ...
+    -------------------------------------
+    311                |  45        |
+    3112               |  10        |
+    311221             |  55        |
+
+    Additionally, this function adds a column called "descendants", which for
+    each industry holds all the descendant industries or industry groups that
+    have detailed information provided in the dataset. After mapping to
+    industries, but before attribution is performed, this column is used by the
+    clean_mapped_mecs_energy_fba function to drop any row that is mapped
+    from an aggregated industry group to a less aggregated industry or industry
+    group THAT HAS DETAILED INFORMATION GIVEN IN THE MECS (and therefore has
+    its own row already) to avoid the over-attribution issue.
+    Again using the previous example:
+
+    ActivityConsumedBy | FlowAmount | descendants | ...
+    ---------------------------------------------------
+    311                |  45        | 3112 311221 |
+    3112               |  10        | 311221      |
+    311221             |  55        |             |
+
+    Note that this function is not useful if the desired aggregation level is
+    NAICS-2. In such a case, the MECS dataset can be filtered to include only
+    the rows with ActivityConsumedBy == "31-33", then disaggregated to 31, 32,
+    33 using another dataset (such as the QCEW).
+    '''
+    mecs = (
+        fba
+        .query('ActivityConsumedBy != "31-33"')
+        .assign(descendants='')
+    )
+
+    for level in [5, 4, 3]:
+        descendants = (
+            mecs
+            .drop(columns='descendants')
+            .query(f'ActivityConsumedBy.str.len() > {level}')
+            .assign(
+                parent=lambda x: x.ActivityConsumedBy.str.slice(stop=level)
+            )
+            .groupby(['Flowable', 'Location', 'parent'])
+            .agg({'FlowAmount': 'sum', 'ActivityConsumedBy': ' '.join})
+            .reset_index()
+            .rename(columns={'ActivityConsumedBy': 'descendants',
+                             'FlowAmount': 'descendant_flows',
+                             'parent': 'ActivityConsumedBy'})
+        )
+
+        mecs = (
+            mecs
+            .merge(descendants,
+                   how='left',
+                   on=['Flowable', 'Location', 'ActivityConsumedBy'],
+                   suffixes=(None, '_y'))
+            .fillna({'descendant_flows': 0, 'descendants_y': ''})
+            .assign(
+                descendants=lambda x: x.descendants.mask(x.descendants == '',
+                                                         x.descendants_y),
+                FlowAmount=lambda x: (x.FlowAmount - x.descendant_flows).mask(
+                    x.FlowAmount - x.descendant_flows < 0, 0)
+            )
+            .drop(columns=['descendant_flows', 'descendants_y'])
+        )
+
+    return mecs
+
+
+def clean_mapped_mecs_energy_fba(fba: FlowByActivity, **_) -> FlowByActivity:
+    '''
+    This function finishes handling the over-attribution issue described in
+    the documentation for clean_mecs_energy_fba by dropping any row in the
+    MECS dataset which has been mapped to an industry or industry group which
+    is a subset (strict or otherwise) of an industry group listed in the
+    descendants columns. So, if 311 and 3112 both appear in the MECS datset,
+    3112 will be listed as a descendant of 311 and this function will therefore
+    drop a row mapping 311 to 311221 (since more detailed information on 3112,
+    which contains 311221, is provided). If 31122 and 311221 do not appear in
+    the dataset, a row mapping 3112 to 311221 will not be dropped, since no
+    more detailed information on 311221 is given. Further attribution/
+    disaggregation should be done using another datatset such as the QCEW.
+    '''
+    mecs = (
+        fba
+        .assign(to_keep=fba.apply(
+            lambda x: not any([str(x.SectorConsumedBy).startswith(d)
+                               for d in x.descendants.split()]),
+            axis='columns'
+        ))
+        .query('to_keep')
+        .drop(columns=['descendants', 'to_keep'])
+    )
+
+    return mecs
+
+
+
+
 def clean_mapped_mecs_energy_fba_to_state(
         fba: FlowByActivity, **_
     ) -> FlowByActivity:
@@ -501,9 +650,6 @@ def mecs_land_fba_cleanup(fba, **_):
     :param fba: df, EIA MECS Land FBA format
     :return: df, EA MECS Land FBA
     """
-    from flowsa.data_source_scripts.temp_data_source_functions import \
-        clean_mecs_energy_fba
-
     fba2 = clean_mecs_energy_fba(fba)
 
     # calculate the land area in addition to building footprint
