+++ flowsa/methods/flowbysectormethods/CAP_HAP_state_2017.yaml
@@ -21,8 +21,8 @@ sources_to_cache:
           - Table 3.2
     exclusion_fields:
       Location: '00000'
-    estimate_suppressed: !script_function:temp_data_source_functions estimate_suppressed_mecs_energy
-    clean_fba: !script_function:temp_data_source_functions clean_mecs_energy_fba
+    estimate_suppressed: !script_function:EIA_MECS estimate_suppressed_mecs_energy
+    clean_fba: !script_function:EIA_MECS clean_mecs_energy_fba
     clean_fba_w_sec: !script_function:EIA_MECS clean_mapped_mecs_energy_fba_to_state
     attribution_method: proportional
     attribution_source:
